components: move ./AuthToken* -> ./security/
