+++ src/App.vue
@@ -21,7 +21,7 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 </template>
 
 <script lang="ts">
-import AuthTokenSection from './components/AuthTokenSection.vue'
+import AuthTokenSection from './components/security/AuthTokenSection.vue'
 import { defineComponent } from 'vue'
 
 export default defineComponent({
