+++ src/components/security/AuthTokenList.vue
@@ -48,7 +48,7 @@
 <script lang="ts">
 import { translate as t } from '@nextcloud/l10n'
 import { defineComponent } from 'vue'
-import { useAuthTokenStore } from '../store/authtoken'
+import { useAuthTokenStore } from '../../store/authtoken'
 
 import AuthToken from './AuthToken.vue'
