+++ src/components/security/AuthTokenSection.vue
